neuralagentAI/desktop/aiagent/requirements.txt
requests
Pillow
pyautogui
mss
pyinstaller
nuitka
pyperclip
httpx
websockets
pywin32
psutil
# pywinsandbox
uiautomation
# pyobjc
# pyatspi